+++ src/main/java/com/siberanka/axbedrockmenus/MenuButton.java
@@ -2,5 +2,8 @@ package com.siberanka.axbedrockmenus;
 
 import java.util.List;
 
-public record MenuButton(String text, List<String> commands, String permission) {
+public record MenuButton(String text, List<String> commands, String permission, List<String> description) {
+    public MenuButton(String text, List<String> commands, String permission) {
+        this(text, commands, permission, List.of());
+    }
 }