+++ src/main/java/com/siberanka/axbedrockmenus/MenuData.java
@@ -2,5 +2,8 @@ package com.siberanka.axbedrockmenus;
 
 import java.util.List;
 
-public record MenuData(String title, List<MenuButton> buttons) {
+public record MenuData(String title, String content, List<MenuButton> buttons) {
+    public MenuData(String title, List<MenuButton> buttons) {
+        this(title, "", buttons);
+    }
 }